refactor: updateProductQuantity, updateAmount 함수의 위치를 productPurchaseModel로 변경
